Move logic to repository layer
