+++ apiresources/MenuOptions.php
@@ -21,7 +21,6 @@ class MenuOptions extends ApiController
             'update' => [],
             'destroy' => [],
         ],
-        'locationAware' => [],
         'request' => Requests\MenuOptionRequest::class,
         'repository' => Repositories\MenuOptionRepository::class,
         'transformer' => Transformers\MenuOptionTransformer::class,
