+++ apiresources/Tables.php
@@ -21,7 +21,6 @@ class Tables extends ApiController
             'update' => [],
             'destroy' => [],
         ],
-        'locationAware' => [],
         'request' => \Admin\Requests\Table::class,
         'repository' => Repositories\TableRepository::class,
         'transformer' => Transformers\TableTransformer::class,
