+++ apiresources/repositories/OrderRepository.php
@@ -8,4 +8,8 @@ use Igniter\Api\Classes\AbstractRepository;
 class OrderRepository extends AbstractRepository
 {
     protected $modelClass = Orders_model::class;
+
+    protected static $locationAwareConfig = [];
+
+    protected static $customerAwareConfig = [];
 }
