+++ apiresources/repositories/ReservationRepository.php
@@ -8,4 +8,8 @@ use Igniter\Api\Classes\AbstractRepository;
 class ReservationRepository extends AbstractRepository
 {
     protected $modelClass = Reservations_model::class;
+
+    protected static $locationAwareConfig = [];
+
+    protected static $customerAwareConfig = [];
 }
