+++ apiresources/repositories/ReviewRepository.php
@@ -8,4 +8,8 @@ use Igniter\Local\Models\Reviews_model;
 class ReviewRepository extends AbstractRepository
 {
     protected $modelClass = Reviews_model::class;
+
+    protected static $locationAwareConfig = [];
+
+    protected static $customerAwareConfig = [];
 }
